Прод: молчит и не пишет в PSRAM; ввод отдаётся раньше запоминания
- G920_TRACE_BYTES: трасса трафика получила свой флаг, на проде не заводится
  и g920_trace_write в образ не попадает
- esp_log_level_set("*", ESP_LOG_NONE) гасит лог самого ESP-IDF; накладка
  SDKCONFIG_DEFAULTS проверена и отвергнута — PlatformIO её не применяет
- в приёмном пути отдача консоли идёт первой, копия последнего отчёта после

Проверено на плате: прод молчит после Calling app_main(), 565 строк -> 0.

+++ firmware/common/util/include/g920/log.h
@@ -130,7 +130,13 @@ void g920_log_hexdump(g920_log_level_t level, const char *tag,
 #define G920_LOG_BUILD_LEVEL G920_LOG_LEVEL_INFO
 #endif
 
-#define G920_LOG_ENABLED(level) ((int)(level) <= (int)(G920_LOG_BUILD_LEVEL))
+/*
+ * Без приведений типа намеренно: обе стороны — целочисленные макросы, и в
+ * таком виде условие годится и для `#if` препроцессора. С `(int)` внутри
+ * `#if` компилятор говорит «missing binary operator before token "("» —
+ * проверено.
+ */
+#define G920_LOG_ENABLED(level) ((level) <= (G920_LOG_BUILD_LEVEL))
 
 #define G920_LOG_AT(level, call)                                              \
     do {                                                                      \
